refactor(domain): Decimal money fields + UTC-aware datetimes
A1 — monetary/price/rate fields migrated from float to decimal.Decimal:
  UserAccount {cash_balance, net_worth, month_start_net_worth};
  LongPosition.avg_entry; ShortPosition {entry_price, locked_cash,
  locked_fund}; PricePoint.price; Stock {current, high_24h, low_24h,
  all_time_high}; HedgeFund.cash_balance + investors values;
  FundPenalty.penalty_apr.

  Exception signatures updated: InsufficientFunds and
  FundInsufficientBalance now accept Decimal. Format spec ":,.2f"
  works natively on Decimal so user-facing messages are byte-identical.

  Quantisation policy documented in the module docstring (currency
  -> Decimal('0.01'), rates -> Decimal('0.0001')). Construction-time
  invariants do not auto-quantise; callers are responsible.

  Duration/count fields (voice_minutes, role_ping_joins,
  role_ping_join_minutes) remain float.

A2 — ActivityBucket.bucket_start default_factory now returns
  datetime.now(tz=UTC). datetime.utcnow is deprecated in Python 3.12
  and returns naive datetimes, which would force naive/aware mixing
  through the Phase 4 persistence boundary.

Resolves the Phase 3.1 blocking items captured in
docs/reviews/pr-11-team-review.md (A1 + A2).

Refs #2.

+++ src/friendex/domain/errors.py
@@ -17,6 +17,7 @@ Derived from ``docs/02-target-architecture.md`` §Error Handling.
 """
 
 from datetime import time
+from decimal import Decimal
 
 
 class DomainError(Exception):
@@ -49,7 +50,7 @@ class DiscordError(FriendexError):
 
 
 class InsufficientFunds(DomainError):
-    def __init__(self, need: float, have: float) -> None:
+    def __init__(self, need: Decimal, have: Decimal) -> None:
         super().__init__(f"Insufficient funds: need ${need:,.2f}, have ${have:,.2f}.")
         self.need = need
         self.have = have
@@ -111,7 +112,7 @@ class InvalidAmount(DomainError):
 
 
 class FundInsufficientBalance(DomainError):
-    def __init__(self, need: float, have: float) -> None:
+    def __init__(self, need: Decimal, have: Decimal) -> None:
         super().__init__(f"Fund balance too low: need ${need:,.2f}, have ${have:,.2f}.")
         self.need = need
         self.have = have
